// src/app/services/openapi-client/model/orderDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OrderItemDto } from './orderItemDto';


export interface OrderDto { 
    id?: number;
    orderNumber?: string | null;
    userId?: number | null;
    status?: string | null;
    paymentStatus?: string | null;
    paymentMethod?: string | null;
    shippingMethod?: string | null;
    subtotal?: number;
    shippingFee?: number;
    discountAmount?: number;
    taxAmount?: number;
    totalAmount?: number;
    shippingFullName?: string | null;
    shippingPhone?: string | null;
    shippingAddress?: string | null;
    items?: Array<OrderItemDto> | null;
    createdAt?: string | null;
    updatedAt?: string | null;
}